# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Synaptiq API specification — Synaptic Applications
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      operationId: getCurrentUser
      summary: Get (and JIT-provision) the current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{id}/role:
    patch:
      tags: [Users]
      operationId: updateUserRole
      summary: Update a user's role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /agent/scrape:
    post:
      tags: [Agent]
      operationId: runScrape
      summary: Scrape a public trending-topic data source (--scrape=<source> --path=<dest>)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeInput"
      responses:
        "201":
          description: Completed scrape job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeJob"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "502":
          description: Upstream source fetch failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /agent/process:
    post:
      tags: [Agent]
      operationId: runProcess
      summary: Process raw scraped data into engineered features (--process=<dest> --features=<csv>)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessInput"
      responses:
        "201":
          description: Completed feature set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureSet"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Scrape job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /agent/train:
    post:
      tags: [Agent]
      operationId: runTrain
      summary: Train the trending-topic agent on a feature set (--train=<csv>)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainInput"
      responses:
        "201":
          description: Completed training run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingRun"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Feature set not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserRole:
      type: string
      enum: [guest, user, admin]
    User:
      type: object
      required: [id, clerkUserId, email, role, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          type: string
          format: date-time
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
...
    JobStatus:
      type: string
      enum: [pending, running, completed, failed]
    ScrapeInput:
      type: object
      required: [dataSourceId]
      properties:
        dataSourceId:
          type: integer
        path:
          type: string
          description: Optional logical destination path/label for the raw dump (mirrors CLI --path=<DEST>).
    ScrapeJob:
      type: object
      required: [id, dataSourceId, dataSourceName, status, recordsScraped, createdAt]
      properties:
        id:
          type: integer
        dataSourceId:
          type: integer
        dataSourceName:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        rawObjectPath:
          type: ["string", "null"]
        recordsScraped:
          type: integer
        errorMessage:
          type: ["string", "null"]
        triggeredByUserId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
    ProcessInput:
      type: object
      required: [scrapeJobId, features]
      properties:
        scrapeJobId:
          type: integer
        features:
          type: array
          items:
            type: string
          description: Requested feature column names (mirrors CLI --features=<CSV>).
    FeatureSet:
      type: object
      required: [id, scrapeJobId, status, featureNames, rowCount, createdAt]
      properties:
        id:
          type: integer
        scrapeJobId:
          type: integer
        status:
          $ref: "#/components/schemas/JobStatus"
        featureNames:
          type: array
          items:
            type: string
        rowCount:
          type: integer
        csvObjectPath:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
    TrainInput:
      type: object
      required: [featureSetId]
      properties:
        featureSetId:
          type: integer
    TrendingTopic:
      type: object
      required: [topic, score]
      properties:
        topic:
          type: string
        score:
          type: number
    TrainingRun:
      type: object
      required: [id, featureSetId, status, createdAt]
      properties:
        id:
          type: integer
        featureSetId:
          type: integer
        status:
          $ref: "#/components/schemas/JobStatus"
        metrics:
          type: ["object", "null"]
          additionalProperties: true
        predictedTopics:
          type: array
          items:
            $ref: "#/components/schemas/TrendingTopic"
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
...